move files closer
