+++ docs/reference/indices/list_query.rst
@@ -1,7 +1,7 @@
 Querying a List Index
 =====================
 
-.. automodule:: gpt_index.indices.query.list
+.. automodule:: gpt_index.indices.list
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
