+++ docs/reference/indices/struct_store_query.rst
@@ -1,7 +1,7 @@
 Querying a Structured Store Index
 =================================
 
-.. automodule:: gpt_index.indices.query.struct_store
+.. automodule:: gpt_index.indices.struct_store
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
