+++ docs/reference/indices/table_query.rst
@@ -1,7 +1,7 @@
 Querying a Keyword Table Index
 ==============================
 
-.. automodule:: gpt_index.indices.query.keyword_table
+.. automodule:: gpt_index.indices.keyword_table
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
