+++ docs/reference/indices/tree_query.rst
@@ -1,7 +1,7 @@
 Querying a Tree Index
 =====================
 
-.. automodule:: gpt_index.indices.query.tree
+.. automodule:: gpt_index.indices.tree
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
