+++ docs/reference/indices/vector_store_query.rst
@@ -7,7 +7,7 @@ query classes specific to each vector store.
 Base Vector Store Query Class
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-.. automodule:: gpt_index.indices.query.vector_store.base
+.. automodule:: gpt_index.indices.vector_store.base
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
@@ -16,7 +16,7 @@ Base Vector Store Query Class
 Vector Store-specific Query Classes
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-.. automodule:: gpt_index.indices.query.vector_store.queries
+.. automodule:: gpt_index.indices.vector_store.queries
    :members:
    :inherited-members:
    :exclude-members: index_struct, query, set_llm_predictor, set_prompt_helper
