+++ gpt_index/indices/empty/__init__.py
@@ -1,7 +1,9 @@
 """Empty Index."""
 
 from gpt_index.indices.empty.base import GPTEmptyIndex
+from gpt_index.indices.empty.query import GPTEmptyIndexQuery
 
 __all__ = [
     "GPTEmptyIndex",
+    "GPTEmptyIndexQuery",
 ]
