+++ gpt_index/indices/empty/base.py
@@ -10,7 +10,7 @@ from typing import Any, Optional, Sequence
 from gpt_index.data_structs.data_structs_v2 import EmptyIndex
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.indices.base import BaseGPTIndex, QueryMap
-from gpt_index.indices.query.empty.base import GPTEmptyIndexQuery
+from gpt_index.indices.empty.query import GPTEmptyIndexQuery
 from gpt_index.indices.query.schema import QueryMode
 from gpt_index.indices.service_context import ServiceContext
