+++ gpt_index/indices/keyword_table/__init__.py
@@ -2,6 +2,11 @@
 
 # indices
 from gpt_index.indices.keyword_table.base import GPTKeywordTableIndex
+from gpt_index.indices.keyword_table.query import (
+    GPTKeywordTableGPTQuery,
+    GPTKeywordTableRAKEQuery,
+    GPTKeywordTableSimpleQuery,
+)
 from gpt_index.indices.keyword_table.rake_base import GPTRAKEKeywordTableIndex
 from gpt_index.indices.keyword_table.simple_base import GPTSimpleKeywordTableIndex
 
@@ -9,4 +14,7 @@ __all__ = [
     "GPTKeywordTableIndex",
     "GPTSimpleKeywordTableIndex",
     "GPTRAKEKeywordTableIndex",
+    "GPTKeywordTableGPTQuery",
+    "GPTKeywordTableRAKEQuery",
+    "GPTKeywordTableSimpleQuery",
 ]
