+++ gpt_index/indices/list/__init__.py
@@ -1,7 +1,11 @@
 """List-based data structures."""
 
 from gpt_index.indices.list.base import GPTListIndex
+from gpt_index.indices.list.embedding_query import GPTListIndexEmbeddingQuery
+from gpt_index.indices.list.query import GPTListIndexQuery
 
 __all__ = [
     "GPTListIndex",
+    "GPTListIndexEmbeddingQuery",
+    "GPTListIndexQuery",
 ]
