+++ gpt_index/indices/list/embedding_query.py
@@ -4,11 +4,9 @@ from typing import Any, List, Optional, Tuple
 
 from gpt_index.data_structs.data_structs_v2 import IndexList
 from gpt_index.data_structs.node_v2 import Node
-from gpt_index.indices.query.embedding_utils import (
-    SimilarityTracker,
-    get_top_k_embeddings,
-)
-from gpt_index.indices.query.list.query import BaseGPTListIndexQuery
+from gpt_index.indices.list.query import BaseGPTListIndexQuery
+from gpt_index.indices.query.embedding_utils import (SimilarityTracker,
+                                                     get_top_k_embeddings)
 from gpt_index.indices.query.schema import QueryBundle
 
 logger = logging.getLogger(__name__)
