--- gpt_index/indices/query/empty/__init__.py
@@ -1,5 +1,0 @@
-"""Query classes for empty indices."""
-
-from gpt_index.indices.query.empty.base import GPTEmptyIndexQuery
-
-__all__ = ["GPTEmptyIndexQuery"]
