--- gpt_index/indices/query/keyword_table/__init__.py
@@ -1,13 +1,0 @@
-"""Query classes for keyword table indices."""
-
-from gpt_index.indices.query.keyword_table.query import (
-    GPTKeywordTableGPTQuery,
-    GPTKeywordTableRAKEQuery,
-    GPTKeywordTableSimpleQuery,
-)
-
-__all__ = [
-    "GPTKeywordTableGPTQuery",
-    "GPTKeywordTableRAKEQuery",
-    "GPTKeywordTableSimpleQuery",
-]
