--- gpt_index/indices/query/list/__init__.py
@@ -1,6 +1,0 @@
-"""Query classes for list indices."""
-
-from gpt_index.indices.query.list.embedding_query import GPTListIndexEmbeddingQuery
-from gpt_index.indices.query.list.query import GPTListIndexQuery
-
-__all__ = ["GPTListIndexEmbeddingQuery", "GPTListIndexQuery"]
