--- gpt_index/indices/query/struct_store/__init__.py
@@ -1,8 +1,0 @@
-"""Query classes for structured store indices."""
-
-from gpt_index.indices.query.struct_store.sql import (
-    GPTNLStructStoreIndexQuery,
-    GPTSQLStructStoreIndexQuery,
-)
-
-__all__ = ["GPTNLStructStoreIndexQuery", "GPTSQLStructStoreIndexQuery"]
