--- gpt_index/indices/query/tree/__init__.py
@@ -1,11 +1,0 @@
-"""Query classes for tree indices."""
-
-from gpt_index.indices.query.tree.embedding_query import GPTTreeIndexEmbeddingQuery
-from gpt_index.indices.query.tree.leaf_query import GPTTreeIndexLeafQuery
-from gpt_index.indices.query.tree.retrieve_query import GPTTreeIndexRetQuery
-
-__all__ = [
-    "GPTTreeIndexLeafQuery",
-    "GPTTreeIndexRetQuery",
-    "GPTTreeIndexEmbeddingQuery",
-]
