--- gpt_index/indices/query/vector_store/__init__.py
@@ -1,5 +1,0 @@
-"""Query classes for vector store indices."""
-
-from gpt_index.indices.query.vector_store.base import GPTVectorStoreIndexQuery
-
-__all__ = ["GPTVectorStoreIndexQuery"]
