+++ gpt_index/indices/struct_store/__init__.py
@@ -1,9 +1,21 @@
 """Structured store indices."""
 
 from gpt_index.indices.struct_store.pandas import GPTPandasIndex
+from gpt_index.indices.struct_store.pandas_query import GPTNLPandasIndexQuery
 from gpt_index.indices.struct_store.sql import (
     GPTSQLStructStoreIndex,
     SQLContextContainerBuilder,
 )
+from gpt_index.indices.struct_store.sql_query import (
+    GPTNLStructStoreIndexQuery,
+    GPTSQLStructStoreIndexQuery,
+)
 
-__all__ = ["GPTSQLStructStoreIndex", "SQLContextContainerBuilder", "GPTPandasIndex"]
+__all__ = [
+    "GPTSQLStructStoreIndex",
+    "SQLContextContainerBuilder",
+    "GPTPandasIndex",
+    "GPTNLStructStoreIndexQuery",
+    "GPTSQLStructStoreIndexQuery",
+    "GPTNLPandasIndexQuery",
+]
