+++ gpt_index/indices/struct_store/pandas.py
@@ -2,14 +2,14 @@
 
 from typing import Any, Optional, Sequence
 
-import pandas as pd
-
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.data_structs.table_v2 import PandasStructTable
 from gpt_index.indices.base import QueryMap
 from gpt_index.indices.query.schema import QueryMode
-from gpt_index.indices.query.struct_store.pandas import GPTNLPandasIndexQuery
 from gpt_index.indices.struct_store.base import BaseGPTStructStoreIndex
+from gpt_index.indices.struct_store.pandas_query import GPTNLPandasIndexQuery
+
+import pandas as pd
 
 
 class GPTPandasIndex(BaseGPTStructStoreIndex[PandasStructTable]):
