+++ gpt_index/indices/struct_store/sql.py
@@ -2,22 +2,21 @@
 import json
 from typing import Any, Dict, Optional, Sequence
 
-from sqlalchemy import Table
-
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.data_structs.table_v2 import SQLStructTable
 from gpt_index.indices.base import BaseGPTIndex, QueryMap
 from gpt_index.indices.common.struct_store.schema import SQLContextContainer
-from gpt_index.indices.common.struct_store.sql import SQLStructDatapointExtractor
+from gpt_index.indices.common.struct_store.sql import \
+    SQLStructDatapointExtractor
 from gpt_index.indices.query.schema import QueryMode
-from gpt_index.indices.query.struct_store.sql import (
-    GPTNLStructStoreIndexQuery,
-    GPTSQLStructStoreIndexQuery,
-)
 from gpt_index.indices.service_context import ServiceContext
 from gpt_index.indices.struct_store.base import BaseGPTStructStoreIndex
-from gpt_index.indices.struct_store.container_builder import SQLContextContainerBuilder
+from gpt_index.indices.struct_store.container_builder import \
+    SQLContextContainerBuilder
+from gpt_index.indices.struct_store.sql_query import (
+    GPTNLStructStoreIndexQuery, GPTSQLStructStoreIndexQuery)
 from gpt_index.langchain_helpers.sql_wrapper import SQLDatabase
+from sqlalchemy import Table
 
 
 class GPTSQLStructStoreIndex(BaseGPTStructStoreIndex[SQLStructTable]):
