+++ gpt_index/indices/tree/__init__.py
@@ -2,7 +2,13 @@
 
 # indices
 from gpt_index.indices.tree.base import GPTTreeIndex
+from gpt_index.indices.tree.embedding_query import GPTTreeIndexEmbeddingQuery
+from gpt_index.indices.tree.leaf_query import GPTTreeIndexLeafQuery
+from gpt_index.indices.tree.retrieve_query import GPTTreeIndexRetQuery
 
 __all__ = [
     "GPTTreeIndex",
+    "GPTTreeIndexLeafQuery",
+    "GPTTreeIndexRetQuery",
+    "GPTTreeIndexEmbeddingQuery",
 ]
