+++ gpt_index/indices/tree/embedding_query.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional, Tuple, cast
 from gpt_index.data_structs.data_structs_v2 import IndexGraph
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.indices.query.schema import QueryBundle
-from gpt_index.indices.query.tree.leaf_query import GPTTreeIndexLeafQuery
+from gpt_index.indices.tree.leaf_query import GPTTreeIndexLeafQuery
 from gpt_index.indices.utils import get_sorted_node_list
 from gpt_index.prompts.prompts import TreeSelectMultiplePrompt, TreeSelectPrompt
