+++ gpt_index/indices/vector_store/__init__.py
@@ -1,6 +1,7 @@
 """Vector-store based data structures."""
 
 from gpt_index.indices.vector_store.base import GPTVectorStoreIndex
+from gpt_index.indices.vector_store.base_query import GPTVectorStoreIndexQuery
 from gpt_index.indices.vector_store.vector_indices import (
     ChatGPTRetrievalPluginIndex,
     GPTChromaIndex,
@@ -22,4 +23,5 @@ __all__ = [
     "GPTChromaIndex",
     "GPTOpensearchIndex",
     "ChatGPTRetrievalPluginIndex",
+    "GPTVectorStoreIndexQuery",
 ]
