+++ gpt_index/indices/vector_store/vector_indices.py
@@ -18,7 +18,7 @@ from gpt_index.data_structs.data_structs_v2 import (
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.indices.base import BaseGPTIndex, QueryMap
 from gpt_index.indices.query.schema import QueryMode
-from gpt_index.indices.query.vector_store.queries import (
+from gpt_index.indices.vector_store.queries import (
     ChatGPTRetrievalPluginQuery,
     GPTChromaIndexQuery,
     GPTFaissIndexQuery,
