+++ tests/indices/embedding/test_base.py
@@ -9,7 +9,7 @@ import pytest
 from gpt_index.data_structs.node_v2 import Node
 from gpt_index.embeddings.base import mean_agg
 from gpt_index.embeddings.openai import OpenAIEmbedding
-from gpt_index.indices.query.tree.embedding_query import GPTTreeIndexEmbeddingQuery
+from gpt_index.indices.tree.embedding_query import GPTTreeIndexEmbeddingQuery
 from gpt_index.indices.tree.base import GPTTreeIndex
 from gpt_index.langchain_helpers.chain_wrapper import (
     LLMChain,
